pyproject.toml
[project]
name = "voice-command-automation"
version = "1.0.0"
description = "Production-ready NLP-based voice command automation system for air-borne systems console"
requires-python = ">=3.10"
authors = [{name = "Auto Control Team"}]
license = {text = "MIT"}

dependencies = [
    # Speech Recognition
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    
    # NLP & Intent Classification
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # CLI & Configuration
    "click>=8.1.7",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    
    # Utilities
    "librosa>=0.10.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.2",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.11",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

[project.scripts]
voice-control = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
